+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]## [0.1.2.2b]
 ### Changed
+- Stop Server button now uses os.kill() instead of deprecated Werkzeug callback
 
 ## [0.1.2.1] - 2002-04-28
 ### Changed